demo_OCR_adaptive_system/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
google-genai>=1.20.0,<2
openai>=1.60.0,<3
qdrant-client>=1.12.1,<2
httpx>=0.27,<0.29
Pillow==10.4.0
requests==2.32.3
gunicorn==22.0.0
